Move SSO tests to standard directory and module layout
